+++ interpreter/jslib/wast.ml
@@ -4,15 +4,15 @@
 open Wasm
 open Js_of_ocaml
 
-let _ =
+let () =
   Js.export "WebAssemblyText"
     (object%js (_self)
 
       method encode (s : Js.js_string Js.t) : (Typed_array.arrayBuffer Js.t) =
-        let def = Parse.string_to_module (Js.to_string s) in
+        let _, def = Parse.Module.parse_string (Js.to_string s) in
         let bs =
           match def.Source.it with
-          | Script.Textual m -> (Encode.encode m)
+          | Script.Textual m -> Encode.encode m
           | Script.Encoded (_, bs) -> bs
           | Script.Quoted (_, _) -> failwith "Unsupported" in
         let buf = new%js Typed_array.arrayBuffer (String.length bs) in