+++ interpreter/runtime/global.ml
@@ -8,7 +8,8 @@ exception Type
 exception NotMutable
 
 let alloc (GlobalT (_mut, t) as ty) v =
-  if not (Match.match_val_type (type_of_value v) t) then raise Type;
+  assert Free.((val_type t).types = Set.empty);
+  if not (Match.match_val_type [] (type_of_value v) t) then raise Type;
   {ty; content = v}
 
 let type_of glob =
@@ -20,5 +21,5 @@ let load glob =
 let store glob v =
   let GlobalT (mut, t) = glob.ty in
   if mut <> Var then raise NotMutable;
-  if not (Match.match_val_type (type_of_value v) t) then raise Type;
+  if not (Match.match_val_type [] (type_of_value v) t) then raise Type;
   glob.content <- v